Cleaning up the file save for tokenizing dataset
The current version of the function does not work well with midi paths with directories with them in python notebooks, this is because a directory has to created before a file can be stored. For example, if the midi path is "a/b/c.mid",  and the output directory is "d/e" then the code gives an error unless "d/e/a/b/" directory exists. This code will automatically take the name of the MIDI file and saves it in the directory as "d/e/c.json". There is a commented code in line 239 if we want to concatenate the entire path as "d/e/a/b/c.mid.json"